refactor: Rename `lib/plugins/aws/package/lib/getHashForFilePath.js`
