+++ lib/plugins/aws/package/compile/functions.js
@@ -8,7 +8,7 @@ const _ = require('lodash');
 const path = require('path');
 const ServerlessError = require('../../../../serverless-error');
 const deepSortObjectByKey = require('../../../../utils/deep-sort-object-by-key');
-const getHashForFilePath = require('../lib/getHashForFilePath');
+const getHashForFilePath = require('../lib/get-hash-for-file-path');
 const parseS3URI = require('../../utils/parse-s3-uri');
 const { log } = require('@serverless/utils/log');
