+++ lib/plugins/aws/package/lib/get-hash-for-file-path.js
